[Silver III] Title: 주유소, Time: 436 ms, Memory: 36332 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,8 +8,8 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
         int N = Integer.parseInt(st.nextToken());
-        int[] len = new int[N-1];
-        int[] gas = new int[N];
+        long[] len = new long[N-1];
+        long[] gas = new long[N];
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N-1; i++) {
             len[i] = Integer.parseInt(st.nextToken());
@@ -18,9 +18,9 @@ public class Main {
         for (int i = 0; i < N; i++) {
             gas[i] = Integer.parseInt(st.nextToken());
         }
-        int pre = gas[0];
-        int charge = 0;
-        int sum = 0;
+        long pre = gas[0];
+        long charge = 0;
+        long sum = 0;
         for (int i = 0; i < N-1; i++) {
             if(pre>gas[i]){
                 pre = gas[i];
